Move conceal e-mail usecase into sub-module
