+++ src/controllers/concealEmail/concealEmail.go
@@ -5,7 +5,7 @@ import (
 	"github.com/halprin/email-conceal/src/context"
 	"github.com/halprin/email-conceal/src/entities"
 	"github.com/halprin/email-conceal/src/external/lib/errors"
-	"github.com/halprin/email-conceal/src/usecases"
+	"github.com/halprin/email-conceal/src/usecases/concealEmail"
 	"log"
 	"net/http"
 )
@@ -46,7 +46,7 @@ func (receiver ConcealEmailController) Add(arguments map[string]interface{}) (in
 		log.Printf("E-mail to conceal with no description = %s", sourceEmail)
 	}
 
-	var concealEmailUsecase usecases.ConcealEmailUsecase
+	var concealEmailUsecase concealEmail.ConcealEmailUsecase
 	applicationContext.Resolve(&concealEmailUsecase)
 	concealedEmail, err := concealEmailUsecase.Add(sourceEmail, description)
 
@@ -95,7 +95,7 @@ func (receiver ConcealEmailController) Delete(arguments map[string]interface{}) 
 
 	log.Println("Conceal E-mail ID to delete =", concealEmailId)
 
-	var concealEmailUsecase usecases.ConcealEmailUsecase
+	var concealEmailUsecase concealEmail.ConcealEmailUsecase
 	applicationContext.Resolve(&concealEmailUsecase)
 	err := concealEmailUsecase.Delete(concealEmailId)
 	if err != nil {
@@ -114,7 +114,7 @@ func (receiver ConcealEmailController) Update(arguments map[string]interface{}) 
 	concealEmailId, _ := arguments["concealEmailId"].(string)
 	description, _ := arguments["description"].(string)
 
-	var concealEmailUsecase usecases.ConcealEmailUsecase
+	var concealEmailUsecase concealEmail.ConcealEmailUsecase
 	applicationContext.Resolve(&concealEmailUsecase)
 
 	var err error
@@ -132,7 +132,7 @@ func (receiver ConcealEmailController) Update(arguments map[string]interface{}) 
 			"error": errorString,
 		}
 		return http.StatusBadRequest, jsonMap
-	} else if errors.As(err, &usecases.ConcealEmailNotExistError{}) {
+	} else if errors.As(err, &concealEmail.ConcealEmailNotExistError{}) {
 		errorString := err.Error()
 		log.Printf(errorString)
 		jsonMap := map[string]string{
