+++ src/usecases/concealEmail/concealEmail_test.go
@@ -1,4 +1,4 @@
-package usecases
+package concealEmail
 
 import (
 	"errors"
